# app/media_types.go
// Code generated by goagen v1.2.0, DO NOT EDIT.
//
// API "Chat API": Application Media Types
//
// Command:
// $ goagen
// --design=github.com/m0a-mystudy/goa-chat/design
// --out=$(GOPATH)/src/github.com/m0a-mystudy/goa-chat
// --version=v1.2.0-dirty

package app

import (
	"github.com/goadesign/goa"
	"time"
	"unicode/utf8"
)

// A account (default view)
//
// Identifier: application/vnd.account+json; view=default
type Account struct {
	// Date of creation
	Created time.Time `form:"created" json:"created" xml:"created"`
	// ID of room
	ID       string `form:"id" json:"id" xml:"id"`
	Password string `form:"password" json:"password" xml:"password"`
}

// Validate validates the Account media type instance.
func (mt *Account) Validate() (err error) {
	if mt.ID == "" {
		err = goa.MergeErrors(err, goa.MissingAttributeError(`response`, "id"))
	}
	if mt.Password == "" {
		err = goa.MergeErrors(err, goa.MissingAttributeError(`response`, "password"))
	}

	return
}

// AccountCollection is the media type for an array of Account (default view)
//
// Identifier: application/vnd.account+json; type=collection; view=default
type AccountCollection []*Account

// Validate validates the AccountCollection media type instance.
func (mt AccountCollection) Validate() (err error) {
	for _, e := range mt {
		if e != nil {
			if err2 := e.Validate(); err2 != nil {
				err = goa.MergeErrors(err, err2)
			}
		}
	}
	return
}

// a google login (default view)
//
// Identifier: application/vnd.login+json; view=default
type Login struct {
	URL string `form:"url" json:"url" xml:"url"`
}

// Validate validates the Login media type instance.
func (mt *Login) Validate() (err error) {
	if mt.URL == "" {
		err = goa.MergeErrors(err, goa.MissingAttributeError(`response`, "url"))
	}
	return
}

// A Message (default view)
//
// Identifier: application/vnd.message+json; view=default
type Message struct {
	AccountID int       `form:"accountID" json:"accountID" xml:"accountID"`
	Body      string    `form:"body" json:"body" xml:"body"`
	PostDate  time.Time `form:"postDate" json:"postDate" xml:"postDate"`
}

// Validate validates the Message media type instance.
func (mt *Message) Validate() (err error) {

	if mt.Body == "" {
		err = goa.MergeErrors(err, goa.MissingAttributeError(`response`, "body"))
	}

	if utf8.RuneCountInString(mt.Body) < 1 {
		err = goa.MergeErrors(err, goa.InvalidLengthError(`response.body`, mt.Body, utf8.RuneCountInString(mt.Body), 1, true))
	}
	if utf8.RuneCountInString(mt.Body) > 400 {
		err = goa.MergeErrors(err, goa.InvalidLengthError(`response.body`, mt.Body, utf8.RuneCountInString(mt.Body), 400, false))
	}
	return
}

// MessageCollection is the media type for an array of Message (default view)
//
// Identifier: application/vnd.message+json; type=collection; view=default
type MessageCollection []*Message

// Validate validates the MessageCollection media type instance.
func (mt MessageCollection) Validate() (err error) {
	for _, e := range mt {
		if e != nil {
			if err2 := e.Validate(); err2 != nil {
				err = goa.MergeErrors(err, err2)
			}
		}
	}
	return
}

// A room (default view)
//
// Identifier: application/vnd.room+json; view=default
type Room struct {
	// Date of creation
	Created *time.Time `form:"created,omitempty" json:"created,omitempty" xml:"created,omitempty"`
	// description of room
	Description string `form:"description" json:"description" xml:"description"`
	// ID of room
	ID *int `form:"id,omitempty" json:"id,omitempty" xml:"id,omitempty"`
	// Name of room
	Name string `form:"name" json:"name" xml:"name"`
}

// Validate validates the Room media type instance.
func (mt *Room) Validate() (err error) {
	if mt.Name == "" {
		err = goa.MergeErrors(err, goa.MissingAttributeError(`response`, "name"))
	}
	if mt.Description == "" {
		err = goa.MergeErrors(err, goa.MissingAttributeError(`response`, "description"))
	}
	if utf8.RuneCountInString(mt.Description) > 400 {
		err = goa.MergeErrors(err, goa.InvalidLengthError(`response.description`, mt.Description, utf8.RuneCountInString(mt.Description), 400, false))
	}
	return
}

// RoomCollection is the media type for an array of Room (default view)
//
// Identifier: application/vnd.room+json; type=collection; view=default
type RoomCollection []*Room

// Validate validates the RoomCollection media type instance.
func (mt RoomCollection) Validate() (err error) {
	for _, e := range mt {
		if e != nil {
			if err2 := e.Validate(); err2 != nil {
				err = goa.MergeErrors(err, err2)
			}
		}
	}
	return
}